requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# LLM Providers
openai>=1.50.0
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.1
structlog>=24.4.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Validation & Serialization
orjson==3.9.12